refactor(utils): split into shared and command-local subdirectories
Resolve the semantic conflation in utils/ flagged by #424 finding #1:
the directory mixed genuine cross-cutting leaves with helpers that are
the guts of a single command, with no structural signal distinguishing
them.

Sub-group accordingly:
- utils/shared/        date-filter, resource-extensions, ignore, validation
- utils/command-local/ open-helpers, mcp-proxy-helpers, watch-constants,
                       search-helpers

The split axis is 'command-agnostic reusable leaf' vs 'command-local',
not import count. command-local helpers stay test-visible (not moved into
commands/**) deliberately: their logic is pure and cross-platform
(per-OS browser/url derivation, wildcard->regex, url normalization) and
is better covered by direct unit tests than by coarser c8() subprocess
tests. The directory name now states the category and the filename names
the owning command.

The layering guard is unaffected: layerOf() classifies by first path
segment, so utils/shared and utils/command-local both remain 'utils'.

Refs #424
