+++ src/commands/incidents.ts
@@ -4,7 +4,7 @@
 
 import { fetchAllPages } from "../core/client.ts";
 import { defineCommand, dryRun } from "../framework/command-framework.ts";
-import { buildDateFilter, parseBetween } from "../utils/date-filter.ts";
+import { buildDateFilter, parseBetween } from "../utils/shared/date-filter.ts";
 
 /**
  * List incidents
