+++ src/utils/shared/resource-extensions.ts
