+++ src/utils/shared/validation.ts
@@ -7,7 +7,7 @@
  * importing the `framework/` layer.
  */
 
-import { getLogger } from "../core/logger.ts";
+import { getLogger } from "../../core/logger.ts";
 
 /**
  * Validate that a positional argument is present and non-empty.
