+++ tests/integration/watch.test.ts
@@ -14,7 +14,7 @@ import {
 import { tmpdir } from "node:os";
 import { join, resolve } from "node:path";
 import { after, before, describe, test } from "node:test";
-import { DEPLOY_COOLDOWN } from "../../src/utils/watch-constants.ts";
+import { DEPLOY_COOLDOWN } from "../../src/utils/command-local/watch-constants.ts";
 import { pollUntil } from "../utils/polling.ts";
 import { startWatchProcess } from "../utils/watch-process.ts";
