+++ tests/unit/mcp-proxy-auth.test.ts
@@ -5,7 +5,7 @@
 import assert from "node:assert";
 import { afterEach, beforeEach, describe, test } from "node:test";
 import type { Logger } from "../../src/core/logger.ts";
-import { createCamundaFetch } from "../../src/utils/mcp-proxy-helpers.ts";
+import { createCamundaFetch } from "../../src/utils/command-local/mcp-proxy-helpers.ts";
 import { makeMockClient, makeMockLogger } from "../utils/mocks.ts";
 
 describe("createCamundaFetch", () => {
